spline-57 refactoring
