+++ core/src/main/scala/za/co/absa/spline/core/transformations/LineageProjectionMerger.scala
@@ -20,8 +20,8 @@ import java.util.UUID
 import java.util.UUID.randomUUID
 
 import za.co.absa.spline.common.transformations.AsyncTransformation
-import za.co.absa.spline.model.expr.{Alias, AttrRef, Expression, TypedExpression}
-import za.co.absa.spline.model.op.{ExpressionAware, Operation, OperationProps, Projection}
+import za.co.absa.spline.model.expr.{Alias, AttrRef, Expression}
+import za.co.absa.spline.model.op.{Operation, OperationProps, Projection}
 import za.co.absa.spline.model.{Attribute, DataLineage}
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -45,66 +45,7 @@ object LineageProjectionMerger extends AsyncTransformation[DataLineage] {
     Future.successful((lineage /: pipeline) ((lin, f) => f(lin)))
 
   private[transformations] def cleanupReferences(lineage: DataLineage): DataLineage = {
-    val operations = lineage.operations
-
-    val datasets = {
-      val datasetIds =
-        (operations.flatMap(_.mainProps.inputs)
-          ++ operations.map(_.mainProps.output)).distinct
-      lineage.datasets.filter(ds => datasetIds.contains(ds.id))
-    }
-
-    val attributes = {
-      val attributeIds = datasets.flatMap(_.schema.attrs).distinct
-      lineage.attributes.filter(attr => attributeIds.contains(attr.id))
-    }
-
-    val dataTypes = {
-      val expressions = operations.flatMap {
-        case op: ExpressionAware => op.expressions
-        case _ => Nil
-      }
-
-      val expressionTypeIds: Set[UUID] = {
-        def traverseAndCollect(accumulator: Set[UUID], expressions: List[Expression]): Set[UUID] = expressions match {
-          case Nil => accumulator
-          case exp :: queue =>
-            val updatedAccumulator = exp match {
-              case tex: TypedExpression => accumulator + tex.dataTypeId
-              case _ => accumulator
-            }
-            val updatedQueue = exp.children.toList ++ queue
-            traverseAndCollect(updatedAccumulator, updatedQueue)
-        }
-
-        traverseAndCollect(Set.empty, expressions.toList)
-      }
-
-      val retainedDataTypes = {
-        val allDataTypesById = lineage.dataTypes.map(dt => dt.id -> dt).toMap
-
-        def traverseAndCollectWithChildren(accumulator: Set[UUID], typeIds: List[UUID]): Set[UUID] = typeIds match {
-          case Nil => accumulator
-          case dtId :: queue => traverseAndCollectWithChildren(
-            accumulator = accumulator + dtId,
-            typeIds = allDataTypesById(dtId).childDataTypeIds.toList ++ queue
-          )
-        }
-
-        val referredTypeIds = expressionTypeIds ++ attributes.map(_.dataTypeId)
-        val retainedTypeIds = traverseAndCollectWithChildren(Set.empty, referredTypeIds.toList)
-
-        allDataTypesById.filterKeys(retainedTypeIds).values
-      }
-
-      retainedDataTypes.toSeq
-    }
-
-    lineage.copy(
-      operations = operations,
-      datasets = datasets,
-      attributes = attributes,
-      dataTypes = dataTypes)
+    lineage.rectified
   }
 
   private[transformations] def mergeProjections(lineage: DataLineage): DataLineage = {
